tích hợp thành công thanh toán VNPay (bản test)

+++ UniCharity_BE/src/main/java/com/UniCharity/UniCharity/services/VNPAYService.java
@@ -19,7 +19,7 @@ import java.util.*;
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 @Slf4j
 public class VNPAYService {
-    public String createOrder(HttpServletRequest request, int amount, String orderInfor, String urlReturn){
+    public String createOrder(HttpServletRequest request, long amount, String orderInfor, String urlReturn){
         String vnp_Version = "2.1.0";
         String vnp_Command = "pay";
         String vnp_TxnRef = VNPAYConfig.getRandomNumber(8);
